pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
authors = ["bogdan <telioko@tutanota.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "3.10.12"
Flask = "^2.3.3"
gunicorn = "^21.2.0"
pip = "^23.2.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.1"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
analyzer = 'page_analyzer.app:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
